Consolidate sqr/cubic and split math.hpp by dependency
`sqr` was defined four times across the library: as `Math<T>::sqr` in
math.hpp and as three separate anonymous-namespace copies, two of which
were `double`-only and so unusable from templated code.

- Add free `ipc::sqr` and `ipc::cubic` to a new `math/scalar_math.hpp`.
- `Math<T>::sqr`/`cubic` forward to them, so the public API is unchanged
  and there is a single implementation.
- Drop the duplicates in normal_collisions.cpp and
  voxel_size_heuristic.cpp.

Split math.hpp along its dependencies so consumers that only want the
scalar helpers do not pay for the rest:

- `math/scalar_math.hpp` -- no includes.
- `math/heaviside_type.hpp` -- `HeavisideType`, `OrientationTypes`.
- `math/math.hpp` -- the parts needing eigen_ext, and math.tpp which
  pulls in the autodiff scalars. It still includes both new headers, so
  existing `#include <ipc/math/math.hpp>` keeps working.

normal_collisions.cpp and voxel_size_heuristic.cpp no longer pull in
TinyAD as a result (verified with -H).
