+++ src/ipc/math/CMakeLists.txt
@@ -1,10 +1,12 @@
 set(SOURCES
+  heaviside_type.hpp
   interval.cpp
   interval.hpp
   math.cpp
   math.hpp
   math.tpp
   morton.hpp
+  scalar_math.hpp
 )
 
 target_sources(ipc_toolkit PRIVATE ${SOURCES})
